Made AI move weighting stable.
Fixed exception and floating point errors caused by a 0 children state. This makes the weighting algorithm work with children generated for actual future moves, not just the computer moves. The weighting algorithm is currently at a fixed depth of 2, but once this is made dynamic, the unstable branch can be merged into master.